chore: reshuffle part2 (#10588)
* extract store

* move prop/init to render

* introduce constants file and move all constants there to reduce cyclic dependencies

* move reactive_import to render

* move bubble_event into render, move (raf) task handing into its own file

* expect-error did hide that one
